# docs/openAPI_document/streamflix_openapi.yaml
openapi: 3.1.0
info:
  title: StreamFlix API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
  /profiles/{profileId}/preferences:
    get:
      tags:
        - preferences
      summary: Get preferences
      description: Get viewing preferences for a profile. Supports JSON, XML, CSV.
      operationId: getPreferences
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Preferences retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - profile belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - preferences
      summary: Update preferences
      description: Update viewing preferences for a profile. Supports JSON, XML, CSV.
      operationId: updatePreferences
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePreferencesRequest'
        required: true
      responses:
        '200':
          description: Preferences updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/PreferencesResponse'
        '400':
          description: Invalid input - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - profile belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /titles/{titleId}/seasons:
    get:
      tags:
        - seasons
      summary: Get seasons
      description: Get all seasons for a series. Supports JSON, XML, CSV.
      operationId: getSeasons
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Seasons retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonResponse'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeasonResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Title not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - seasons
      summary: Create season
      description: Create a new season for a series. Supports JSON, XML, CSV.
      operationId: createSeason
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSeasonRequest'
        required: true
      responses:
        '201':
          description: Season created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/SeasonResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/SeasonResponse'
        '400':
          description: Invalid input - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Title not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    UpdatePreferencesRequest:
      properties:
        preferredGenres:
          type: array
          items:
            type: string
            enum:
              - ACTION
              - COMEDY
              - DRAMA
              - HORROR
              - THRILLER
              - ROMANCE
              - DOCUMENTARY
              - ANIMATION
              - SCIENCE_FICTION
              - FANTASY
              - CRIME
              - MYSTERY
          uniqueItems: true
        preferredType:
          type: string
          enum:
            - MOVIE
            - SERIES
        minimumMaturityRating:
          type: string
          enum:
            - ALL
            - PG
            - TEEN
            - MATURE
    PreferencesResponse:
      properties:
        profileId:
          type: string
        preferredGenres:
          type: array
          items:
            type: string
            enum:
              - ACTION
              - COMEDY
              - DRAMA
              - HORROR
              - THRILLER
              - ROMANCE
              - DOCUMENTARY
              - ANIMATION
              - SCIENCE_FICTION
              - FANTASY
              - CRIME
              - MYSTERY
          uniqueItems: true
        preferredType:
          type: string
          enum:
            - MOVIE
            - SERIES
        minimumMaturityRating:
          type: string
          enum:
            - ALL
            - PG
            - TEEN
            - MATURE
...
    CreateSeasonRequest:
      properties:
        seasonNumber:
          type: integer
          format: int32
          maximum: 10000
          minimum: 1
    SeasonResponse:
      properties:
        id:
          type: string
        seasonNumber:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT